requirements.txt
fastapi
google-cloud-bigquery
google-cloud-pubsub
pydantic
python-dotenv
uvicorn
